@@ -1,56 +1,51 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <RuntimeIdentifiers>win-x64;osx-x64;linux-x64</RuntimeIdentifiers>
+    <!--If you are willing to use Windows/MacOS native APIs you will need to create 3 projects.
+    One for Windows with net8.0-windows TFM, one for MacOS with net8.0-macos and one with net8.0 TFM for Linux.-->
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
-    <UseAppHost>true</UseAppHost>
-    <!--Avalonia doesen't support TrimMode=link currently,but we are working on that https://github.com/AvaloniaUI/Avalonia/issues/6892 -->
-    <TrimMode>copyused</TrimMode>
     <BuiltInComInteropSupport>true</BuiltInComInteropSupport>
-    <RootNamespace>Olfactometer</RootNamespace>
-    <BaseOutputPath>..\bin</BaseOutputPath>
-    <Version>1.1.0</Version>
+    <ApplicationManifest>app.manifest</ApplicationManifest>
+    <OutputPath>..\bin\$(Configuration)</OutputPath>
+    <Version>$(AppVersion)</Version>
+    <ApplicationIcon>..\Harp.Olfactometer.Design\Assets\cf-logo.ico</ApplicationIcon>
+    <Name>Harp.Olfactometer.App</Name>
   </PropertyGroup>
 
   <PropertyGroup>
-    <CFBundleName>Olfactometer</CFBundleName>
-    <CFBundleDisplayName>Olfactometer</CFBundleDisplayName>
+    <CFBundleName>Harp.Olfactometer.App</CFBundleName>
+    <CFBundleDisplayName>Harp.Olfactometer.App</CFBundleDisplayName>
     <CFBundleIdentifier>org.fchampalimaud</CFBundleIdentifier>
     <CFBundleVersion>$(Version)</CFBundleVersion>
     <CFBundleShortVersionString>$(Version)</CFBundleShortVersionString>
     <CFBundlePackageType>AAPL</CFBundlePackageType>
     <CFBundleSignature>.</CFBundleSignature>
-    <CFBundleExecutable>Olfactometer</CFBundleExecutable>
+    <CFBundleExecutable>Harp.Olfactometer.App</CFBundleExecutable>
     <CFBundleIconFile>cf-logo.icns</CFBundleIconFile>
     <NSPrincipalClass>NSApplication</NSPrincipalClass>
     <NSHighResolutionCapable>true</NSHighResolutionCapable>
     <Copyright>Champalimaud Foundation</Copyright>
-    <ApplicationIcon>..\Assets\cf-logo.ico</ApplicationIcon>
+    <ApplicationIcon>..\Harp.Olfactometer.Design\Assets\cf-logo.ico</ApplicationIcon>
     <PackageReadmeFile>..\README.md</PackageReadmeFile>
     <RepositoryType>git</RepositoryType>
-    <BaseOutputPath>..\bin</BaseOutputPath>
+    <BaseOutputPath>..\bin\$(Configuration)</BaseOutputPath>
   </PropertyGroup>
 
-  <Target Name="CopyCustomContent" AfterTargets="AfterBuild" Condition="$([MSBuild]::IsOsPlatform('OSX'))">
-    <Copy SourceFiles="..\Assets\cf-logo.icns" DestinationFolder="$(OutDir)"/>
+  <Target Name="CopyCustomContent" AfterTargets="AfterBuild" Condition="$([MSBuild]::IsOsPlatform('OSX')) or $([MSBuild]::IsOsPlatform('Linux'))">
+    <Copy SourceFiles="..\Harp.Olfactometer.Design\Assets\cf-logo.icns" DestinationFolder="$(OutDir)" />
   </Target>
+
   <Target Name="CopyCustomContent2" AfterTargets="Publish">
-    <Copy SourceFiles="..\Assets\cf-logo.ico" DestinationFolder="$(OutDir)\publish"/>
+    <Copy SourceFiles="..\Harp.Olfactometer.Design\Assets\cf-logo.ico" DestinationFolder="$(OutDir)\publish" />
   </Target>
 
   <ItemGroup>
-    <ProjectReference Include="..\Harp.Olfactometer.Design\Harp.Olfactometer.Design.csproj"/>
+    <PackageReference Include="Avalonia.Desktop" Version="$(AvaloniaVersion)" />
+    <PackageReference Include="Dotnet.Bundle" Version="*" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Avalonia" Version="0.10.21"/>
-    <PackageReference Include="Avalonia.Desktop" Version="0.10.21"/>
-    <!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
-    <PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" Version="0.10.21" />
-    <PackageReference Include="DotNet.Bundle" Version="0.9.13"/>
-    <PackageReference Include="XamlNameReferenceGenerator" Version="1.6.1"/>
+    <ProjectReference Include="..\Harp.Olfactometer.Design\Harp.Olfactometer.Design.csproj" />
   </ItemGroup>
-
 </Project>
