@@ -1,72 +1,29 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <RuntimeIdentifiers>win-x64;osx-x64;linux-x64</RuntimeIdentifiers>
-    <RootNamespace>Olfactometer.Design</RootNamespace>
-    <Version>1.1.0</Version>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-    <OutputPath>../bin/Debug/</OutputPath>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-    <OutputPath>../bin/Release/</OutputPath>
+    <LangVersion>latest</LangVersion>
+    <AvaloniaUseCompiledBindingsByDefault>true</AvaloniaUseCompiledBindingsByDefault>
+    <OutputPath>..\bin\$(Configuration)</OutputPath>
+    <Version>$(AppVersion)</Version>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Update="**\*.xaml.cs">
-      <DependentUpon>%(Filename)</DependentUpon>
-    </Compile>
-    <AvaloniaResource Include="**\*.xaml">
-      <SubType>Designer</SubType>
-    </AvaloniaResource>
-    <AvaloniaResource Include="Assets\**"/>
-    <Compile Update="App.axaml.cs">
-      <DependentUpon>App.axaml</DependentUpon>
-      <SubType>Code</SubType>
-    </Compile>
+    <AvaloniaResource Include="Assets\**" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectCapability Include="Avalonia"/>
-    <TrimmerRootAssembly Include="Avalonia.Themes.Fluent"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-Bold.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-BoldItalic.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-ExtraBold.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-ExtraBoldItalic.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-Italic.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-Light.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-LightItalic.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-Regular.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-SemiBold.ttf"/>
-    <EmbeddedResource Include="Assets\Fonts\OpenSans-SemiBoldItalic.ttf"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Avalonia" Version="0.10.21"/>
-    <PackageReference Include="Avalonia.Desktop" Version="0.10.21"/>
+    <PackageReference Include="Avalonia" Version="$(AvaloniaVersion)" />
+    <PackageReference Include="Avalonia.Controls.ColorPicker" Version="$(AvaloniaVersion)" />
+    <PackageReference Include="Avalonia.Fonts.Inter" Version="$(AvaloniaVersion)" />
+    <PackageReference Include="Avalonia.ReactiveUI" Version="$(AvaloniaVersion)" />
+    <PackageReference Include="Avalonia.Themes.Fluent" Version="$(AvaloniaVersion)" />
     <!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
-    <PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" Version="0.10.21"/>
-    <PackageReference Include="Avalonia.ReactiveUI" Version="0.10.21"/>
-    <PackageReference Include="Avalonia.ReactiveUI.Events" Version="0.10.21"/>
-    <PackageReference Include="Avalonia.Svg.Skia" Version="0.10.18"/>
-    <PackageReference Include="HexIO" Version="3.0.5"/>
-    <PackageReference Include="MessageBox.Avalonia" Version="2.1.0"/>
-    <PackageReference Include="ReactiveUI.Fody" Version="19.2.1"/>
-    <PackageReference Include="ReactiveUI.Validation" Version="3.1.7"/>
-    <PackageReference Include="System.IO.Ports" Version="6.0.0"/>
-    <PackageReference Include="XamlNameReferenceGenerator" Version="1.6.1"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\Interface\Harp.Olfactometer\Harp.Olfactometer.csproj"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Models\"/>
+    <PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" Version="$(AvaloniaVersion)" />
+    <PackageReference Include="HeroIcons.Avalonia" Version="1.0.3" />
+    <PackageReference Include="Markdown.Avalonia" Version="11.0.2" />
+    <PackageReference Include="MessageBox.Avalonia" Version="3.2.0" />
+    <PackageReference Include="ReactiveUI.Fody" Version="19.5.41" />
+    <ProjectReference Include="..\..\Interface\Harp.Olfactometer\Harp.Olfactometer.csproj" />
   </ItemGroup>
 </Project>
