@@ -1,10 +1,12 @@
-﻿using Avalonia;
+﻿using System;
+using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
-using Olfactometer.Design.ViewModels;
-using Olfactometer.Design.Views;
 
-namespace Olfactometer.Design;
+using Harp.Olfactometer.Design.ViewModels;
+using Harp.Olfactometer.Design.Views;
+
+namespace Harp.Olfactometer.Design;
 
 public partial class App : Application
 {
@@ -19,10 +21,24 @@ public partial class App : Application
         {
             desktop.MainWindow = new MainWindow
             {
-                DataContext = new MainWindowViewModel(),
+                DataContext = new OlfactometerViewModel()
+            };
+        }
+        else if (ApplicationLifetime is ISingleViewApplicationLifetime singleViewPlatform)
+        {
+            singleViewPlatform.MainView = new OlfactometerView
+            {
+                DataContext = new OlfactometerViewModel()
             };
         }
 
         base.OnFrameworkInitializationCompleted();
+    }
+    
+    private void NativeMenuItem_OnClick(object sender, EventArgs e)
+    { 
+        var about = new About() { DataContext = new AboutViewModel() };
+        about.ShowDialog((Application.Current.ApplicationLifetime as IClassicDesktopStyleApplicationLifetime)
+            .MainWindow);
     }
 }
